backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.7",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
